# backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Mahreen Indonesia API",
    "description": "Dokumentasi API backend Mahreen Indonesia (Express 5 + MySQL). Endpoint admin memerlukan JWT Bearer token; refresh token memakai HttpOnly cookie.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/api/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "http",
      "scheme": "bearer",
      "bearerFormat": "JWT",
      "description": "Ketik token akses JWT Anda (Bearer <token>)."
    }
  },
  "paths": {
...
    "/api/webinars/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/webinars/{slug}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/webinars/{slug}/registrations": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "x-requested-with",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "fullName": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "whatsapp": {
                  "example": "any"
                },
                "institution": {
                  "example": "any"
                },
                "profession": {
                  "example": "any"
                },
                "city": {
                  "example": "any"
                },
                "webinarSlug": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/webinars/{slug}/payments": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "registrationId": {
                  "example": "any"
                },
                "method": {
                  "example": "any"
                },
                "bank": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/api/admin/newsroom/webinars": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "slug": {
                  "example": "any"
                },
                "title": {
                  "example": "any"
                },
                "category": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "duration": {
                  "example": "any"
                },
                "price": {
                  "example": "any"
                },
                "isFree": {
                  "example": "any"
                },
                "image": {
                  "example": "any"
                },
                "scheduleDate": {
                  "example": "any"
                },
                "scheduleTime": {
                  "example": "any"
                },
                "topics": {
                  "example": "any"
                },
                "mentors": {
                  "example": "any"
                },
                "timeline": {
                  "example": "any"
                },
                "benefits": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                },
                "quota": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/admin/newsroom/webinars/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "force",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/api/client/newsroom/webinars": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "type": "string"
          },
          {
            "name": "category",
            "in": "query",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/newsroom/webinars": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "type": "string"
          },
          {
            "name": "category",
            "in": "query",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "slug": {
                  "example": "any"
                },
                "title": {
                  "example": "any"
                },
                "category": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "duration": {
                  "example": "any"
                },
                "price": {
                  "example": "any"
                },
                "isFree": {
                  "example": "any"
                },
                "image": {
                  "example": "any"
                },
                "scheduleDate": {
                  "example": "any"
                },
                "scheduleTime": {
                  "example": "any"
                },
                "topics": {
                  "example": "any"
                },
                "mentors": {
                  "example": "any"
                },
                "timeline": {
                  "example": "any"
                },
                "benefits": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                },
                "quota": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/newsroom/webinars/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "force",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  },
...
}